Match some outline positioning behavior from RGSS and enable transparent outlines
Outlines should 1) Not change the position of the text and 2) Be visible even if the text is cut off from reaching the edge of the rect, without leaving the rect itself.
We're currently not drawing full text for the outlines like RGSS does, so we don't perfectly succeed at (2), but if we ever start doing that then the logic is already there.

Text transparency with outlines present still doesn't match Enterbrain's implementation, but I gave up on trying to figure that out.